+++ examples/nostr-shell-example-application/src/main/java/org/tbk/nostr/example/shell/command/IdentityCommand.java
@@ -8,7 +8,7 @@ import org.springframework.shell.standard.ShellMethod;
 import org.tbk.nostr.example.shell.util.Json;
 import org.tbk.nostr.identity.Identity;
 import org.tbk.nostr.identity.MoreIdentities;
-import org.tbk.nostr.nips.Nip19;
+import org.tbk.nostr.nip19.Nip19;
 
 import java.io.IOException;